refactor(encoding): Phase 1e Stage C batch 14 — migrate LDM helpers onto impl BtMatcher
Move all four `ldm_*` helpers (`ldm_skip_raw_seq_store_bytes`,
`ldm_get_next_match_and_update_seq_store`, `ldm_maybe_add_match`,
`ldm_process_match_candidate`) off `impl HcMatchGenerator` onto
`impl BtMatcher` — the LDM seq store sits in `BtMatcher::ldm_sequences`
and the helpers only read `MatchCandidate` / mutate the caller-owned
`HcOptLdmState`, so no `MatchTable` borrow is needed.

Macro `build_optimal_plan_impl!` (call sites at the macro definitions
of `start_matching_optimal` body) and the one HcMatchGenerator-side
caller updated to `\$self.bt.ldm_*` / `self.bt.ldm_*`.

Build (default + --no-default-features), strict clippy, 381 lib tests
incl. level22 ratio gate — all green.
